LPS-161984 Extract AttachmentFormBase from objectFieldFormBase to its own component
